backend/requirements.txt
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-decouple==3.8
supabase==2.4.2
django-filter==23.5
django-environ==0.13.0